backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "1.0.0"
description = "Backend for the AI-powered chatbot with lead capture functionality"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "firebase-admin>=6.2.0",
    "google-generativeai>=0.3.1",
    "pydantic>=2.5.0",
]

[project.scripts]
start = "startup:start_server"
